+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Add class to assign users a device ID cookie and provide the value for logging
 * Add class to wrap accessing / setting / deleting cookies for injectability and testability
 * Add helper method to get a DateTimeImmutable from a unix timestamp, in current timezone